[WAL-87] Moved several activity logic to fragment as it was not needed to be on activity
