+++ app/src/main/java/com/asfoundation/wallet/ui/backup/BackupActivityPresenter.kt
@@ -3,10 +3,8 @@ package com.asfoundation.wallet.ui.backup
 import android.content.Context
 import android.content.Intent
 import androidx.documentfile.provider.DocumentFile
-import com.asfoundation.wallet.billing.analytics.WalletsAnalytics
 import com.asfoundation.wallet.billing.analytics.WalletsEventSender
 import com.asfoundation.wallet.ui.BaseActivity
-import com.asfoundation.wallet.ui.backup.creation.BackupCreationFragment
 
 class BackupActivityPresenter(private val view: BackupActivityView,
                               private val data: BackupActivityData,
@@ -18,17 +16,7 @@ class BackupActivityPresenter(private val view: BackupActivityView,
     if (isCreating) navigator.showBackupScreen(data.walletAddress)
   }
 
-  fun sendWalletSaveFileEvent() {
-    when (view.getCurrentFragment()) {
-      BackupCreationFragment::class.java.simpleName -> {
-        eventSender.sendWalletSaveFileEvent(WalletsAnalytics.ACTION_BACK,
-            WalletsAnalytics.STATUS_FAIL, WalletsAnalytics.REASON_CANCELED)
-      }
-    }
-  }
-
-  fun onActivityResult(requestCode: Int, resultCode: Int,
-                       data: Intent?, context: Context) {
+  fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?, context: Context) {
     if (requestCode == BackupActivity.ACTION_OPEN_DOCUMENT_TREE_REQUEST_CODE) {
       var systemFileIntentResult = SystemFileIntentResult()
       if (resultCode == BaseActivity.RESULT_OK && data != null) {
