+++ app/src/main/java/com/asfoundation/wallet/ui/backup/BackupActivityView.kt
@@ -4,15 +4,9 @@ import io.reactivex.Observable
 
 interface BackupActivityView {
 
-  fun askForWritePermissions()
-
   fun closeScreen()
 
-  fun onPermissionGiven(): Observable<Unit>
-
   fun onSystemFileIntentResult(): Observable<SystemFileIntentResult>
-
-  fun getCurrentFragment(): String
 
   fun setupToolbar()
